# specs/001-todo-frontend-auth/contracts/todo-api-contract.yaml
openapi: 3.0.0
info:
  title: Todo API for Frontend Authentication
  version: 1.0.0
  description: API contract for todo management operations accessible to authenticated users

servers:
  - url: https://api.example.com/v1
    description: Production server
  - url: http://localhost:8000/v1
    description: Development server

paths:
  /auth/login:
    post:
      summary: Authenticate user and return JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  description: User's email address
                password:
                  type: string
                  format: password
                  description: User's password
      responses:
        '200':
          description: Successful authentication
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT token for subsequent requests
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials

  /auth/logout:
    post:
      summary: Logout user and invalidate session
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successfully logged out

  /todos:
    get:
      summary: Get all todos for the authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of user's todos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
        '401':
          description: Unauthorized
    post:
      summary: Create a new todo for the authenticated user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoInput'
      responses:
        '201':
          description: Todo created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '401':
          description: Unauthorized

  /todos/{id}:
    get:
      summary: Get a specific todo by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Todo ID
      responses:
        '200':
          description: Todo details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '401':
          description: Unauthorized
        '404':
          description: Todo not found
    put:
      summary: Update a specific todo by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Todo ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoInput'
      responses:
        '200':
          description: Todo updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '401':
          description: Unauthorized
        '404':
          description: Todo not found
    delete:
      summary: Delete a specific todo by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Todo ID
      responses:
        '204':
          description: Todo deleted successfully
        '401':
          description: Unauthorized
        '404':
          description: Todo not found

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          description: Unique user identifier
        email:
          type: string
          format: email
          description: User's email address
        name:
          type: string
          description: User's display name
        createdAt:
          type: string
          format: date-time
          description: User account creation timestamp

    Todo:
      type: object
      properties:
        id:
          type: string
          description: Unique todo identifier
        title:
          type: string
          description: Todo title
        description:
          type: string
          description: Todo description
        completed:
          type: boolean
          description: Completion status
        userId:
          type: string
          description: Owner user ID
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp

    TodoInput:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          maxLength: 200
          description: Todo title (1-200 characters)
        description:
          type: string
          maxLength: 1000
          description: Todo description (0-1000 characters)
        completed:
          type: boolean
          description: Completion status